Rename bluetooth_page package to main_page
